backend-2/requirements.txt
Flask==2.3.2
flask-cors
python-dotenv
